try double distance and vCenter in kmeans

+++ src/main/java/li/tengfei/apng/opt/optimizer/ColorUtils.java
@@ -13,7 +13,7 @@ public class ColorUtils {
     /**
      * color distance
      */
-    public static int distance(Color a, Color b) {
+    public static double distance(Color a, Color b) {
         int dist = 0;
         int delta = a.getRed() - b.getRed();
         dist += delta * delta;
@@ -23,6 +23,7 @@ public class ColorUtils {
         dist += delta * delta;
         delta = a.getAlpha() - b.getAlpha();
         dist += delta * delta;
+        //dist *= Math.sqrt(a.getAlpha() + b.getAlpha());
         return dist;
     }
